Huge rework of the human-package, drippling through almost all other components
The model was updated to be much more on point, which resulted in
changes necessary in every component that used its features. This were
basically all. To keep the code compiling and the tests succeeding, one
big commit is used, instead of many small ones. These changes should be
treated as a unit (it makes no sense to look at them bit by bit as
well).
Necessary UI updates are not included, and will follow.