Solve conflicts for attribute persistence
Due to the introduction of the new and shiny attribute converters using
the `@Lob` annotation on the same attributes crashes hibernate. It is
unclear what persistence strategy should actually be used (serialized
because of `@Lob` or through the attribute converter). Therefore
removing the `@Lob` annotations as the converters should be used.