Rework filter infrastructure
Each filter will now be created depending on the kind of objects it
wants to filter and how the filter itself will be created - e.g. a
filter does not have to perform the filter itself but may rather create
an object which can be used for filtering afterwards.

To get even more practical: the filter as it was implemented so far was
expected to receive some kind of Stream<T> and would the perform the
filtering on this stream, resulting in some potentially smaller amount
of results.
Now it is also possible to fill a Predicate which will contain all the
filter criteria. In the end this predicate will be returned and may then
be used by some Spring Repository to do some database-level filtering.
This scenario was also the driving factor for reworking the whole
infrastructure.