Overhaul event architecture in person module
Instead of creating a new instance each time some data should be updated
the update(...) methods will now operate directly on their respective
objects.
This mainly solves an issue with the distribution of domain events by
Spring: when updating a persisted entity, the new data will be merged.
As this attached entity does not contain the domain events, none will be
published.
By fixing these problems the overall API becomes a lot clearer as well:
1. no more calls like
`personManager.updatePerson(person.getParticipantProfile().updateDateOfBirth(updatedDob))`
2. when creating an updated person, it would potentially share some
events with the original one
3. the Edit...FormExtractor services become much more service-like,
updating their repective objects instead of returning a new instance
of how the object would look like. This in part is a first contribution
to refining the manager, controller and services

See: https://stackoverflow.com/a/48042661/5161760
This solves part of the issues described in #30. The rework of the
managers still remains to be done however.