#33 - Fix JPA annotations on PhoneNumber
To force the format to be applied to all instances, property-based access will be used for now.
Furthermore the column name now matches the name used in the Person class for easier migration. Last but not least some query methods had to be marked as transient to prevent them from being persisted.