#30 - Publish first draft of reworked person overview template
Still work in progress. *Breaks existing functionality*. Do not apply until development has finished.
Most notable: the `PersonController` with all its related controllers will be moved to a separate package. The same will apply to the controllers for events, etc. Also instead of delivering huge HTML-views, functionality will be split up into different pages. This also enables the usage of more narrowly-focused controllers as well as the inclusion of fragments in the Thymleaf-templates.