#30 - Restructure template design of the person profile page; add referent profile
The restructuring was necessary because with lazy-loading the validation states of forms could not be retained and passed through the various controllers easily (i.e. without hacking). The new layout basically placed all tabs in separate templates and used full-fledged controllers to control them. So instead of always loading `/persons/$id/` separate pages like `/persons/$id/activist-referent/` are displayed.
This is achived by using separate controllers for each page which put the necessary data into the model. The `personDetails.html` then dispatches on the template that should actually be used and includes the corresponding fragment. This keeps both the controllers small as well as the templates clear. On the downside some "artificial" controllers have to be introduced to take care of the rendering of each tab (e.g. the `ActivistReferentController` for the activist/referent profile tab).