Add infrastructure to use the 'add person' form as an intermediate dialog
This adds inputs which capture an action as well as redirection to use after a person has been successfully added. Instead of redirecting to the person's detail page as the controller would normally do, it will now check if a redirection parameter exists and if that is the case use this one instead. The action will be provided as a GET parameter to be analyzed by the resulting template.