Fix unintuitive display of time spans
In case of nights' stay the displayed time span would start at the first
day (which is equivalent to the first night), but end at the date of the
last overnight stay. This is contrary to common sense which also
includes the day after the last overnight stay (i.e. the day of
departure).